+++ tests/test_hashRouting.js
@@ -63,6 +63,32 @@ test("location delegate", function() {
   ok(onChangeCallback.wasCalled);
   ok(mockLocation.replace.wasCalled);
   equal(mockLocation.replace.mostRecentCall.args[0], '#/hello_replace_test');
+
+  /**
+    * Test silent assign
+    */
+  onChangeCallback.reset();
+  mockLocation.assign.reset();
+
+  var request = new Davis.Request({ fullPath: '/hello_assign_silent_test', method: 'get' });
+  locationDelegate.assign(request, {silent: true});
+
+  ok(!onChangeCallback.wasCalled);
+  ok(mockLocation.assign.wasCalled);
+  equal(mockLocation.assign.mostRecentCall.args[0], '#/hello_assign_silent_test');
+
+  /**
+    * Test replace
+    */
+  onChangeCallback.reset();
+  mockLocation.assign.reset();
+
+  var request = new Davis.Request({ fullPath: '/hello_replace_test', method: 'get' });
+  locationDelegate.replace(request, {silent: true});
+
+  ok(!onChangeCallback.wasCalled);
+  ok(mockLocation.replace.wasCalled);
+  equal(mockLocation.replace.mostRecentCall.args[0], '#/hello_replace_test');
 });
 
 test("normalizing the initial value of window.location", function() {